+++ packages/ranger-core/src/apollo/adobeLink.ts
@@ -1,8 +1,11 @@
 import { ApolloLink, HttpLink } from '@apollo/client'
 import { onError } from '@apollo/client/link/error'
 import { shrinkFetchQuery } from '@ranger-theme/apollo'
-import { parseCookies } from '@ranger-theme/utils'
+import { events, parseCookies, removeItem } from '@ranger-theme/utils'
 import { isEmpty } from 'lodash-es'
+
+import { logURLLink } from './logURL'
+import { validApiMessage } from './validApi'
 
 export type AdobeLinkType = {
   cookies: any
@@ -19,8 +22,9 @@ export const createAdobeLink = (config: AdobeLinkType) => {
   if (isVercelDeploy) apiURL = `${config.domain}/`
 
   const suffix: string = isAdobeDeploy ? 'graphql' : 'api/graphql'
+  const uri: string = `${typeof window === 'undefined' ? `${apiURL}` : `${window.location.origin}/`}${suffix}`
   const httpLink = new HttpLink({
-    uri: `${typeof window === 'undefined' ? `${apiURL}` : `${window.location.origin}/`}${suffix}`,
+    uri,
     credentials: 'same-origin',
     fetch: shrinkFetchQuery as any,
     useGETForQueries: true
@@ -48,12 +52,22 @@ export const createAdobeLink = (config: AdobeLinkType) => {
   const errorLink = onError(({ graphQLErrors, networkError }: any) => {
     if (graphQLErrors) {
       graphQLErrors.forEach(({ message }: any, index: number) => {
-        graphQLErrors[index].message = message.replace('GraphQL error: ', '')
+        const msg: string = message.replace('GraphQL error: ', '')
+        graphQLErrors[index].message = msg
 
         if (typeof window !== 'undefined') {
-          window.notification.warning({
-            message: graphQLErrors[index].message
-          })
+          const validFlag: boolean = validApiMessage(msg)
+
+          if (validFlag) {
+            removeItem(null, 'access_token')
+            removeItem(null, 'cart_id')
+            removeItem(null, 'compare_uid')
+            events.emit('handleLogout')
+          } else {
+            window.notification.warning({
+              message: graphQLErrors[index].message
+            })
+          }
         } else {
           console.error(graphQLErrors[index].message)
         }
@@ -72,5 +86,6 @@ export const createAdobeLink = (config: AdobeLinkType) => {
   })
 
   const apolloLink = middlewareLink.concat(httpLink)
-  return errorLink.concat(apolloLink)
+  const finalLink = errorLink.concat(apolloLink)
+  return ApolloLink.from([logURLLink(uri), finalLink])
 }