+++ packages/ranger-hooks/src/timedown/useTimedown.ts
@@ -1,22 +1,18 @@
 import { useRef, useState } from 'react'
 
-export const useCountdown = (delay: number = 1000) => {
+export const useTimedown = (delay: number = 1000) => {
   const callbackRef = useRef<any>(null)
   const [count, setCount] = useState<number>(0)
   const [timer, setTimer] = useState<any>(null)
   const [disabled, setDisabled] = useState<boolean>(false)
 
-  if (count === 0) {
-    if (timer) clearInterval(timer)
-    if (timer) setDisabled(false)
-    if (timer) setTimer(null)
-  }
-
   callbackRef.current = () => {
-    setCount((prev: number) => prev - 1)
+    if (count > 0) setCount((prev: number) => prev - 1)
   }
 
-  const handleCountdown = (value: number) => {
+  const handleTimedown = (value: number) => {
+    if (timer) clearInterval(timer)
+
     const time: any = setInterval(() => {
       callbackRef.current()
     }, delay)
@@ -26,9 +22,15 @@ export const useCountdown = (delay: number = 1000) => {
     setTimer(time)
   }
 
+  if (count === 0) {
+    if (timer) clearInterval(timer)
+    if (timer) setDisabled(false)
+    if (timer) setTimer(null)
+  }
+
   return {
     disabled,
     count,
-    handleCountdown
+    handleTimedown
   }
 }